Move moisture-related functions to new utils header
