+++ Source/Utils/Make.package
@@ -12,7 +12,7 @@ CEXE_headers += ERF_TileNoZ.H
 CEXE_headers += ERF_Utils.H
 CEXE_headers += ERF_EpochTime.H
 CEXE_headers += ERF_ParFunctions.H
-
+CEXE_headers += ERF_MoistUtils.H
 CEXE_headers += ERF_SatMethods.H
 CEXE_headers += ERF_WaterVaporSaturation.H
 CEXE_headers += ERF_DirectionSelector.H
